Refactor to useOffline hook
